Add files via upload
So, this is what I have so far. What I'd like to do is have it take a screenshot that has a different name every time, like, 001, 002, 003, etc. I can't seem to get it to work though. I'd also like the screenshot to be taken on a button press, not on a key having to be pressed.